Add files via upload
It is an program in Cpp to create a simple number analysis tool for a basic
application.It needs to perform operations on a list of numbers provided by
the user. Specifically, it is need to compute the sum, average, and count of even and
odd numbers from the list.It will use loops to process these numbers.